Allow names given in unix or RISC OS format to be converted.
If invoked from the RISC OS AMU file, the filenames will be in RISC OS
format, and we'll want to be able to process them as if we were on
RISC OS. So we try to convert the filenames to the format that would
be valid if we were on RISC OS - which means converting them from the
non-native format to the native format. Only the RISC OS to non-RISC OS
path is implemented (that's the one that will be used most of the time).