Created a skeleton document; fixed a few rendering issues.
There is now a skeleton document that can be generated using the
riscos-prminxml tool. This is just copied from the catalog.

A number of rendering issues, with things like the error numbers
on error definitions not being on newlines, links not appearing
properly in related sections, declaration not being rendered at
all, and updated some dates.