Add variable expansion to RISC OS paths; allow XML file copying.
We can now cause the XML file to be copied alongside the generated
HTML, which means that we will be able to distribute them in a form
which allows them to be edited. Yay.

RISC OS paths are now expanded on non-RISC OS systems. This is a
little tedious but necessary because of the way that the riscos-amu
system passes on the filenames from the Makefile system.