Add in a HowTo guide; add examples to system vars; replace development.
There is now a HowTo guide to talk you through how to use the system.
There are examples in the System variables. The 'development' label has
now gone from the 1.02 version, and I've removed the remaining movspclr
references from all the 1.02 documents.

All the documentation has moved into a 'docs' directory now, so that it's
a little easier to find.