Add report on failures when the index build fails.
The build would not report failures usefully when the index failed -
it only reported 'look in this file', which isn't very helpful in
a CI reporting environment.

We now report the log file with a little indentation to show which
bits are the blocks, and which bits are the actual messages. The
validation that follows was not reporting any form of output, which
meant you could be completely oblivious to the fact that it didn't
pass the lint phase.

We now report a count of the failures when linting (but still don't
make this a terminal condition).