Add the ability to check the files with linting.
Linting is now an option that can be applied before all others, so
that we know whether there are any problems with the files. The
default behaviour is to not lint the files, but it would be good
if we could allow the reports to be given as a failure linting did
not pass.