Initial version of the HTML5/CSS based transformation.
This is some initial work on the stylesheet for the HTML5/CSS styled
transformation. It covers the root element and some of the simpler
elements so far. It's at least enough to handle the basic header
and the section text. Services are nearly handled as well.

The CSS should be easy enough to apply across the board, but it will
still need quite a bit of checking to make sure that the transform
is correct for different documents.