Refactored and simplified the syntax on definitions.
The syntax on definitions has now been simplified to make it easier
to extend. The *-command version has been used as the base as it's
the most featureful, and the VDU version now uses the common template.

This means that the VDU syntax will look a little different in 1.03
to 1.02 because the rules have changed slightly on what the content
should be. However, there is only one chapter that uses the VDU
definitions, and it can be changed easily.

Both the HTML and HTML5 transformations have been updated to reflect
this change.