Updated the styles for many of the remaining elements.
The remaining elements are now handled as a simple set of spans
where possible, and pre where not. Some of the CSS has been updated
to use 'rem' instead of 'em' so that we do sizing based on the
root instead of relative to the variable size of the content, which
gets confusing.