Initial work to refactor the definition header into a common template.
The definition header is largely the same on each of the types of
definition with variations due to the original PRM format and what
attributes make sense. In many cases this can be reduced to the same
problem, and the common parts of these definitions made more obviously
the same by being expressed in a single called template.

This change updetes the SWI and vector headers so that they use the
new format. There are more wide ranging changes needed for the others
due to the name and number prefixes.