Replace all the definition headers with template calls in HTML 5.
The definition headers were all inline with the definitions. This is
not the way to do it and remain maintainable. They have all been
converted to a call to a template which does the work - this avoids
us having lots of problems with the way that the different forms work
differently.