Improvements to the tool for running on RISC OS.
Lots more fixes for the RISC OS side of the execution so that we
can run it either as if it was installed for linux, or (potentially)
installed within !Catalogs. Still not sure the best way to do it,
but probably installing in !Catalogs is best.

Most of the fixes require the xsltproc and xmllint versions that
I've just released - 1.41 - because they use the fixed versions of
the --native handling, and catalogs.