Rework the stylesheet handlign to allow variants of the base CSS.
The stylesheet handling has been improved to allow a variant to be
selected, which will be added to the base CSS. This allows the
chapters to be generated in a much more structured form in the
future, taking from external CSS files (not yet implemented).

The 'acornfs' functional specification has been added which allows
us to select a style that looks like the functional specifications
that were put out by Acorn towards the end. This is largely to
demonstrate the features. The CSS has been tweaked and improved
so that it's easier to see that the tables work properly and there
have been some fixes for existing elements which were not rendering.