Bring the main HTML5 style closer to the PRM; create dedicated PRM variant.
The main HTML5 style is now closer to the PRM, using sans-serif for
the heading blocks in line with the original. There is a new variant
of the traditional format which should make everything look even more
like the PRM, and handles pagination with Prince XML. That makes it
quite neat to produce workable PDFs from the source documents without
too much extra effort.

There are still some problems with it - the bullets seem to come out
really spaced out for some reason, and it's looking like the definition
headings are way more spaced than they were in the original PRM
version. These should all be fixable.