Make the spacing on the tables a little wider.
The spacing on the tables is a little tight in the default presentation.
The HTML and HTML 5 transformation in 1.03 have been improved to make it
easier to view these tables more easily. The HTML 5 version is better,
not increasing the vertical sizing as much, and making the horizontal
spacing clear but not obtrusive.