Improve the page break and divider rendering in HTML 5 prm variant.
When the PRM variant is used this introduces the paging attributes
that break the page up usefully. This might be better in the main
style, but for now it's here. This change makes the dividers vanish
when they would duplicate the divider at the top of the page for the
header, and makes the contents go away in printed form.