Improve the formatting of the headings to match original PRM.
The headings and the positioning is now closer to the size and style
of the PRM in the standard and 'prm' variants of traditional.
This should mean it's a lot less likely to be disliked, I hope.

The ordered lists have been made more sensibly placed now in the
prm variant, so that they don't overhang the indent level.