Add option to supply parameters to the stylesheet.
Parameters are used to control some of the features in the stylesheet
but these were never exposed to the `riscos-prminxml` tool. These are
now supported through the `--param` option which allows them to be
configured at invocation time.

It is likely that the main use of this will be to change the CSS
configuration of the files.