Tidy up table spacing in some cases; improve styling of meta.
The meta section is now styled a little better, with more classes
differentiating parts of the content. This also lead to trying to
reduce the additional vertical spacing of tables within table cells
for the related content. I think this has largely been addressed
by changing the border-spacing to be horizontal only, not vertical,
and then applying padding to the cells where necessary.

A new example 'meta' document has been included which shows some
example content so that it's possible to see how larger blocks will
look.