Add some initial styles to use improved fonts in HTML 5.
Ideally the stylesheets should degrade nicely. However, that's
not always the case, and regularly we'll see that there requirements
from one font which conflict with others. The combination that has
been settled on here allows the use of free fonts or the paid for ones
as fall back.

This works reasonably well to select the fonts I would like to use,
but isn't ideal for all cases. The current changes are a compromise
which moves from ITC Novarese through Fraunces to serif for the
serif font, and from Raleway to sans-serif for the sans-serif
fonts.