Rename 'traditional' to 'standard'; make paged media capable.
The default css style is now called 'standard'. This is more what
it is, so it's a good idea to make it clear what it does. Similarly,
the paged media selectors have been moved into the standard style
from the 'prm' variant. This makes prm simpler, and means that the
common styling in the standard style is now able to handle paging.