Fix for HTML content being rendered wrongly.
Switching to the XML generation but processing the document as HTML
did not work well. It's likely that some special cases will be needed
for the generation into EPub in the future, but right now just
reverting to the HTML output is much more reliable.

In particular this was causing problems for the HTML parsing because
elements like <span .../> were being interpreted as an open span,
which was then implicitly closed when the parent element was closed.
This meant that some elements moved inside empty elements, which
messed up the styling badly.