Update indexed generation to allow more control over page content.
The page content in the indexed generation is now able to handle
passing through the CSS configuration so that we can generate the
correct CSS files according to the requirements. There's still a
problem with the CSS file which will end up not being referenced
properly because the directory it lives in will be different to
that of the actual HTML content. This may need some thought to
generate the correct relative locations.