Rework the expansion of the document content in the contents.
New options have been added to 'include-sections' in the contents
which was previously just an extra index for 'sections' (which wasn't
the right way to do it). This means that when enabled, the sections
in the documentation are expanded into rows on the main contents
page. The depth to which this occurs can be controlled by using the
'include-sections-depth' setting, which limits the sections to only
the depth given.

This has been updated in both the HTML and HTML5 variants for some
consistency. The result is a pretty neat effect in the contents
page.