Add the ability to use Fraunces and Raleway as webfonts.
The intention here is that we have named fonts that can be installed,
and you have the option of taking them as web fonts. This ensures that
when processed as part of an automated system, the results aren't
variable based on google's serving of the fonts, or the network
conditions.