Added a style similar to the RISC OS 2 PRMs.
The RISC OS 2 PRMs had a quite distinctive style with the vertical
dividing like separating the headings from the body content. This
was one of the challenging parts of the re-styling of the content.
It's still not perfect; some parts are styled with body font when
they should be heading font, and vice-versa. But this could easily
be addressed in a manually injected style.