Improve handling of image sizes in HTML5; improve contents in ro2.
The image sizes overlapped the edges of the containers. This looked
ugly, so they are now constrained in HTML5. The image container is
now a figure element, which is more in keeping with the style we
use, and it should mean that we have a much better chance of styling
properly. The captions appear in a 'figcaption' and are aligned, so
that they are centred with respect to the image.

The contents in the prm-ro2 variant looked very ugly; these have
now been fixed so that they look more like the rest of the document's
style.