Try to keep headings with content more in HTML5 printed.
When the HTML 5 version is printed it's useful to ensure that it
keeps the content for the headings so that we don't get widows
on the prior page. This doesn't appear to be that effective for
the Firefox and Chrome printing, but seems to work in PrinceXML.
I'm expecting I've missed something subtle.