Update contents and references in HTML 5 to be consistent.
The contents is now consistent with the references and should look
identical. The definitions use a similar but not exactly the same
set of parameters to allow them to be generated in a way more
in line with the original PRMs. Essentially that means that we don't
reproduce the SWI on the name of the SWI calls in the definition,
but everywhere else we do.