Refactor the references processing in HTML 5 to remove duplicate code.
Each of the definitions, and each of the sections have exactly the
same code, with just the elements that are selected varying. This change
removes that commonality to make it much simpler to add new processing
for all the references of a given type.

The intent is that this will make it easier to add the 'use-reasonname'
attribute, and potentially to read the information from other documents
so that those lookups will work across documents (when a href is
supplied).

It also makes it easier to add in a 'always include numbers in the
references' if we wanted.