Fix for indexed entries on the contents page of the RO2 style.
The index entries weren't appearing as an indented section with the
dividing bar between them. This meant that they looked very ugly
compared to the other parts of the RO2 style.

This change also moves the footer inside the divider section, although
without a header beside it.

We also change the page borders slightly to give a little more
printable area on the inside edge.