Experimental document parameters for the docgroup/part/chapter.
Making it possible to pass through document parameters to override
the style.

The docgroup and chapter number are now overridable. The part number
has been added, and is also overridable. They're not used for anything
yet beyond the original use, so docgroup updates the edge index, and
the title. Chapter updates the title in PRM mode, and part number is
currently ignored.