Move the mouse icons to a variant CSS; improve action labelling.
The mouse icons are now a variant CSS, rather than being in the main
style. This means that they should be easiser for people to change
and they don't end up in the default documents, as they're a
presentational effect beyond the basic.

The action labelling uses the localdb as well now. That should make
it easier to change the text as necessary, and allows the use of the
suffixes for scrolling up/down.