Add warning when the XML_CATALOG_FILES for our resource do not exist.
Alan's seeing an odd problem where the system is failing to set the
XML_CATALOG_FILES environment variable. The only way that should happen
is if the catalog files don't exist.

This change introduces a warning that is displayed when the problem
happens.

Debian has been added to the platforms we test.