Add a 'data-size' field to the offset-table and message-table.
To make it easier to represent the data structures that don't stick
to just word-length fields, a new 'data-size' field has been added.
This should allow us to better represent some of the fields like
the Wimp message blocks and some filing system and network structures
that are otherwise difficult to explain.