Prevent the definition's 'use' content being split on to a new page.
The 'section' for definitions in the CSS is marked as being a block
to keep together. This is ok for the smaller sections but the 'use'
section is large and this commonly means that the bit that you want
to read has been forced on to a separate page. This is undesireable.

The change here updates just the 'use' section, to not avoid the breaks
and instead, makes the first paragraph avoid breaks before (which
will keep it with the 'use' heading' and additionally makes that first
paragraph avoid breaks. This should cause it to pull the use heading
on to a new page if the first paragraph would be too long - otherwise
the section will break as normal.