Add docgroup name and chapter numbers to indexed generation.
The index.xml can now include an explicit docgroup-name in the
options, to allow it to group into something other than the
'Programmers Reference Manuals'. And the chapter number can be
generated and passed to each chapter so that we end up with the
documents being generated with numbers like the originals.