Fix for descriptive paragraphs before definitions not breaking.
The definitions like swi-definition are commonly inside a bare section
that contains the definitions, so the first definition doesn't break
the page. However, sometimes we want to include some descriptive text
before the definitions. This happens very rarely but where it does it
is very useful to ensure that we do force a break in that case, to keep
things tidy.