Fix for the bit element repeating the name of fields with state.
When fields have state, we go out of our way to avoid repeating the
bit number, but then we do repeat the bit name. This looks repetetive
and is unnecessary. We can remove the repeated names so that they are
follow the same rules as the number, IF the name for the bit is the
same.