Initial work to clean up index pages.
The index pages weren't very tidy, and they didn't follow the same
style as the PRM that closely. They can now be much closer to the
original PRMs, or include more information. They subreasons can be
shown in separate lines so that they don't wrap uncomfortably.
The page numbers now appear right aligned, and in theory the resulting
pages are easier to style.