Experiment with chapter-front-matter; reset page numbers at contents.
Experimented with named page types to allow the footer page number
styling to change. This was not successful because although the footer
page number can change, the links to those pages do not - they would
still be styled as a regular set of page numbers. So instad of seeing
a link to preface on page iii, you'd get a link to preface page 3.

The Contents page now resets the page number to 1. This might not be
quite right, but it will do for now - in the future this might be
reversed, so that the page number on the page and the page number
of the document look the same.