Add support for document cover pages in indexed content.
The cover page can be specified with some references. The content
is generated automatically. This is really only intended for simple
title pages where the user wants to supply a cover image and some
simple text. We should provide a more advanced form for when you
wish to supply a HTML file which is pre-formatted.