Support for epigraph front-matter. Documentation for front-matter/covers.
We can now include an epigraph page in the front-matter pages. This
allows quotes to be displayed as part of the document before the
contents.

The elements for front-matter and covers are now included in the
documentation for the indexed document. This remains a part of the
tool, which doesn't feel very nice, but that's what I have right now.