Improved compatibility/version elements.
The compatibility block is now limited to only being used within the
`*-definition` blocks. A new version-table element is used in the other
places, in a similar way to the other `*-table` blocks.

The 'state' elements have been refactored to make them easier to use
consistently. This should mean that we're not going to get different
text because we only changed one place. However, there's still quite a
bit of inconsistency in the structures.