Add plain HTML implementation of compatibility/version elements.
The plain HTML implementation has now been provided. This isn't great,
but it kinda matches the style of the default form of the table in
the HTML 5 version, and still keeps to using the <dl><dt><dd> elements
that we use in the HTML version.

More real world examples added to the document to try to show how it
would be used in reality.