Ensure the GitHub RISC OS build job doesn't run to 6 hr time out.
The GitHub actions have a timeout of 6 hours. Something about the
RISC OS Build client which talks to the service has failure mode
that means that it never progresses the connection. Whilst this
is a problem, it's also useful to limit our running of the job,
as there's no reason for it to run for long periods - it cannot
usefully take more than 10 minutes at present.