Stop defaulting to 'monospace' if font isn't known.
The idea with reverting to Monospace was that if the font was missing
we'd notice really clearly. But it causes a very ugly jump in the style
of the font whilst the page is loading in the browser, and it doesn't
really gain us much. So I'm reverting back to just the 'serif' or
'sans-serif' fonts for those.