Include the reason name in the headings for definitions.
If there is a reason name given, the headings would not include it -
only the links in the Contents would include the reason name. This
made it very difficult to find what they're called. The reason name
is now included on the page - this is a difference from how it was
in the earlier manuals, but it's an important piece of information.