Fix for index sorting with reason codes and offsets.
Reason code numbers had been sorted by their alphabetic string.
This wasn't very nice as 10 appeared before 2. We now sort properly
with the reason number being dealt with as a number.

The offsets and offset-base on the SWIs are now handled properly
in the index - they appear at the start, but they do show up with
proper numbers displayed now, so that we're not left with an empty
'&' as the only value in the number field. This could be tidied up
more, but right now, it is at least better looking.