Add 'Chapter #' to the start of the PDF TOC chapters.
In the PDF the table of contents (bookmarks) now adds the chapter
number, if chapter numbers have been supplied. This makes the
contents a bit easier to navigate if you know the chapter number
(which, of course, you might not, as this is a dynamic book). It
could always be removed in the future, or by CSS if we don't like
it.