Added a new style variant for the PRM - 'modern'.
This modern variant has a number of features that differentiate it
from the older PRM style. The intention is to look more modern,
to avoid the excessive whitespace, and to use a little bit of
colour (green) to highlight the section headings (as most PDFs will
be viewed in the browser, not printed).

* No heavy indentation of content.
* Green text for headings.
* Thinner solid black lines instead of thick grey lines for dividers.
* Tables with borders between rows and columns.
* Definitions listed in the page footers.