Update documentation to be clearer about installation on RISC OS.
The installation on RISC OS was not very clear. I believed it was
obvious what you'd do - but that's the thing about making things
available to other people... they don't have the assumptions you
do! So I've made it clearer how it should be installed and what
should be done with it.