Update the default generation format to HTML 5.
It doesn't really make sense to have the old HTML transformation
be the default. All modern browsers support the CSS required, and
whilst it might not work as well on older RISC OS browsers, that's
something that needs to be addressed independantly. It remains
possible to use the old transformation, but you have to be
explicit about wanting to use that format now.

The examples generated are now clearer what they are generated
from.