Fix for `processor-flag` being disallowed in `entry` elements.
The `processor-flag` wasn't allowed in the `entry` elements because
it's weird - we shouldn't really be encouraging such things, but
it's supported for the vector and SWI entries so that CnP can work.
Nobody else should be relying on the flags like that, but in order
that it be possible to document those interfaces, it's now allowed.